+++ src/main/java/content/ContentType.java
@@ -1,0 +1,16 @@
+package content;
+
+/**
+ * The [ContentType] class...
+ */
+public class ContentType {
+  public static final String URL = "URL";
+
+  public static final String LINK = "LINK";
+
+  public static final String REG = "REG";
+
+  public static final String FILE = "FILE";
+
+  public static final String LRN = "LRN";
+}